make some python exercises compatible with our new test suite
The code in charge of extracting the solution of an exercise for the
test is a bit more fragile than the real templating code. So make the
exercises less surprising for that code.

The right approach may be to fix the test code, but this one is easier
and harmless.