Reduce the dependency of Lecture onto Play
We still depend on play.api.libs.json, and this should be replaced by
another library, such as the one already used in PLM.

The ultimate goal is to move the lessons to their own repository, so
that they can be properly shared between the judge, the server and the
javaUI (when revived).

The current version is not satisfactory because the exercises are
serialized in json between the server and the judge. That's really
inefficient, and even error prone. Right now, asking the judge to
deserialize a turtle world freezes it (pb to be investiguated).

The exercises cannot move back into PLM itself (nor in a fixed jar)
because at some point we want to revive the online editor:
https://github.com/BuggleInc/PLM/pull/453
The dream is that each teacher can assemble a lesson from existing
resources.  This will probably create forks on lessons and exercises,
but if the material is in a git, that's no big deal. Then, the teacher
points the pupils onto the exact commit of the lesson, and every parts
get the material from the git, in the right version using the git hash.