make descriptions in Lesson immutable (var -> val), and non-optional
It was probably motivated to get the Lesson built incrementally, with
most fields coming from the json, and the description then loaded
separately.

But instead, a builder (such as introduced in this commit) gives us
the incremental building without the mutable uglyness (by using a
curried function). That's an imperative monster for a functional
beast. But at least, it seems more idiomatic in Scala :)