:art: fiddle with the overall exercise player for better screen space usage
Don't have a upper and lower part in the window, but directly put the
world change icons below the IDE (in the same column) and the exercise
result under the world viewer (in the same column).

This allows the user to do a better usage of her screen by choosing
when she wants to see which part.

Much more work is needed to improve the look and feel of this
application.